requirements.txt
torch
torchvision
matplotlib
imageio
scikit-image
scipy
cython
torchsummary
pyyaml
attrdict
numpy
opencv-python
tensorboardX
tensorflow
moviepy